Merge Ankur Feature branch with Master

+++ src/main/java/com/example/restservice/crud/Child.java
@@ -13,6 +13,8 @@ import javax.persistence.Table;
 @Table(name = "ChildBasic")
 public class Child {
 
+	//Name, DOB, Mother Tongue, Admission date, Religion, Home code
+	//firstNameAndlastNameAnddateOfBirthAndmotherTongueAndadmissionDateAndReligionAndrainbowHomeNumber
 	private Long childNo;
 	private String childStringId;
 	private Integer rainbowHomeNumber;
@@ -71,7 +73,7 @@ public class Child {
 	
 	private Integer bloodGroup;
 	
-	private Integer aadharNumber;
+	private Long aadharNumber;
 	private String abuseOnStreet;
 	private String noOfTimesLeavingHome;
 	
@@ -357,11 +359,11 @@ public class Child {
 		this.picture = picture;
 	}
 
-	public Integer getAadharNumber() {
+	public Long getAadharNumber() {
 		return aadharNumber;
 	}
 
-	public void setAadharNumber(Integer aadharNumber) {
+	public void setAadharNumber(Long aadharNumber) {
 		this.aadharNumber = aadharNumber;
 	}
